adaptive_ai/scripts/requirements.txt
# Dependencies for the model extraction script
aiohttp>=3.9.0
click>=8.1.0
PyYAML>=6.0.1
tenacity>=8.2.0